Remove unused React hooks
- Remove useOnGlobalEscPress (only used in removed popover component)
- Inline useShowToastForSlackConnectStatus (single usage)
